Work-in-progress FFmpeg guide for all platforms
Please let me know what should be changed, in terms of format or content.

It looks a little disjointed as only the Windows section has been completed.
The wording and formatting are quite rough as this is the initial pass.

Design decisions:
The reader selects the appropriate Downloading section for their operating system.
At the end of each Downloading section there is an internal reference to take the reader to the corresponding Installation section.

Pictures are added to illustrate both the Downloading and Installation sections.

If you are a Mac or Linux user, please complete the other sections.