New FX Docs example
The idea behind this is to create a more organized FX documentation area.

The folder and page structure are based on the node categories in the FX Schematic.

I have included an example page to show the basic idea of layout.
Formatting, styling, image sizes, and content are not intended to be final.